// src/types/ExtendedProduct.ts
export type ProductDescription = {
  title: string;
  text: string[];
};

export type ExtendedProduct = {
  id: string;
  category: string;
  namespaceId: string;
  name: string;
  capacityAvailable: string[];
  capacity: string;
  priceRegular: number;
  priceDiscount: number;
  colorsAvailable: string[];
  color: string;
  images: string[];
  description: ProductDescription[];
  screen: string;
  resolution: string;
  processor: string;
  ram: string;
  cell: string[];
  zoom: string;
  camera: string;
};